+++ config/application.rb
@@ -10,8 +10,9 @@ module TheCraftCentralIntegratedConsignmentSystem
   class Application < Rails::Application
     # Initialize configuration defaults for originally generated Rails version.
     config.load_defaults 5.1
-    config.active_record.default_timezone = :local
-    config.time_zone = 'Singapore'
+    # config.active_record.default_timezone = :local
+    # config.time_zone = 'Singapore'
+    # Groupdate.time_zone = false
     # Settings in config/environments/* take precedence over those specified here.
     # Application configuration should go into files in config/initializers
     # -- all .rb files in that directory are automatically loaded.