+++ db/migrate/20180212084304_create_ending_inventories.rb
@@ -2,6 +2,8 @@ class CreateEndingInventories < ActiveRecord::Migration[5.1]
   def change
     create_table :ending_inventories do |t|
       t.references :store, foreign_key: true
+      t.decimal :sales, precision: 2
+      t.decimal :cogs, precision: 2
 
       t.timestamps
     end